Created a new limited duration application
The limited duration version of the application allows you to use a cron job or task scheduler to run the container only during the monitoring windows.

build.bat maintains the original Dockerfile
build-limited-duration.bat creates a version of the container for a single window. It also uses environment variables to get some of the settings.

This approach can be extended to Kubernetes cron jobs.